Use listeners instead of cloud
The cloud interface provides some pooling/scaling intelligence that
we don't need as that functionality is provided by nodepool.

This branch uses a QueueListener to listen for queued executions,
requesting nodes from nodepool when an appropriate label is seen.

Pipeline jobs can use multiple nodes, and they can be allocated
and released at any point in the job. Consequently, listening to
build completion events is not sufficient for cleanup up nodes.
Instead an executor listener is used to release nodes as soon
as they are deallocated by the Pipeline. This also works for
freestyle jobs.