Refactor around NodePool instance
This commit creates a NodePool object that holds all information
relating to a NodePool connection. This means that multiple connections
can be handled, but it also gets rid of a mess of statics. A NodePools
class is added to contain the NodePool instances.

Classes that were very thing wrappers (such as ZooKeeperClient) have
been removed.

The UI now has a test connection button and some validation.