Suspend agent programmatically
After a task completes, suspend an agent programmatically.  This is the
intended approach to distinguish an agent marking itself as
"unavailable" versus simply going offline due to connectivity issues.
In the latter case, Jenkins will try to re-connect the agent.

This also collapses the ExecutionListener into the NodePoolComputer
class, which already implements the interface through the SlaveComputer
superclass.