Add ability to hold a node for debugging
Add the ability to hold a node for debugging.  A node in the "hold"
state will be retained by NodePool until it is deleted manually.

A node can be held by clicking the checkbox on the following UI page:

"Manage Jenkins" => "Manage Nodes" => Click the gear icon on the
node to be held.