RE-1502 Tighten regex for specific nodes
There is some code in NodePoolQueueListener that attempts to detect
when a build is requesting a specific node rather than a general label
the idea being that builds should never request a specific node, so
those builds will be cancelled.

However we introduced general labels that end with a number and
legitimate jobs were being cancelled.

This patch tightens up the regex so that only labels that end with
dash then ten digits will be caught as nodepool ids are 10 digits long.

Ideally we would get a list of labels from nodepool and compare
agianst that instead of pattern matching, but I can't see a way
of doing that.