RE-2230 Cleanup Deadlock
The Janitor thread initiates a clean every 60s. If a node is found that
needs to be cleaned up, a thread is created to clean it up. This thread
waits for the cleanup lock before proceeding. Once the lock is acquired
it does the cleanup, which invovles creating another thread.

The problem is that these cleanup threads can build up, especially
when there are many nodes to remove. If they build up to the near the
thread limit, then deadlock can occur as one thread has the lock, but
it can't create a thread to complete it's task as there are too many
other threads waiting for the lock.

This commit also adds a timeout to a blocking call in
NodepoolComputer.disconnect so that it can't hold the cleanup lock
indefinitely.