RE-2397 Remove Nodepools.cleanupLock
This lock is acquired inconsistently leading to a deadlock where
one thread holds the Queue lock but needs cleanup lock, and another
holads cleanup but needs the Queue lock.

After considering the code and the consequences, I've decided its best
to remove the cleanup lock. This may result in some KazooLock exceptions
in the logs if two threads attempt to release the same node, but thats
better than causing a Jenkins master hang due to deadlock. This is the
second deadlock bug thats been found relating to this cleanup lock and
I don't want to discover any more.